Attempt at LDA/STA and a generic result container.
The write back container is so that we don't have to worry about where
we're writing to (so it could be A register or some memory). It's not
entirely complete as I'm fairly sure I got some of the destinations
wrong in my grid. Also it's not entirely complete simply because STA for
example can't be writen wbc_ = data_ just yet because I'd almost need to
duplicate the instToAddressMode array so that one just gets A in data_
and the other one uses some other thing for indexFetch(). It would be a
total pain to have yet another 256 enum just because it's not great
looking but hey...